requirements.txt
django-phonenumber-field
phonenumbers
django == 2.2.7